refactor(core): move registry to core; add plugin module for cmip6; clean processing package; fix lint
