--- src/piddiplatsch/processing/__init__.py
@@ -1,2 +1,0 @@
-# Processing package no longer exposes BaseProcessor.
-__all__: list[str] = []
