+++ tests/test_processor_registry.py
@@ -1,7 +1,7 @@
 """Test to verify processor registry extensibility."""
 
 from piddiplatsch.core.processing import BaseProcessor
-from piddiplatsch.processing.registry import (
+from piddiplatsch.core.registry import (
     get_processor,
     list_processors,
     register_processor,
